seems ok now, to further test
