+++ sphinxsim/sph_simulation/continuum_simulation_builder.cpp
@@ -83,7 +83,8 @@ void ContinuumSimulationBuilder::buildSimulation(SPHSimulation &sim, const json 
     //----------------------------------------------------------------------
     // Define basic state recording for visualization the simulation results.
     //----------------------------------------------------------------------
-    auto &body_state_recorder = createBodyStatesRecording(sph_system, config_manager, main_methods, config);
+    auto &body_state_recorder = io_builder_ptr_->createBodyStatesRecording(
+        sph_system, config_manager, main_methods, config);
     //----------------------------------------------------------------------
     //	Define time-integration method, screen out uput and observation sample rate.
     //----------------------------------------------------------------------
